Move script component tests to script helper tests
